pyproject.toml
[project]
name = "specify-x"
version = "0.1.0"
description = "specKitEnhanced — opinionated SDD framework wrapping github/spec-kit"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Shai Elchuk", email = "shailelchuk@gmail.com" }]

dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    "jsonschema>=4.22",
    "python-frontmatter>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
specify-x = "specify_x.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specify_x"]
